requirements.txt
transformers==4.46.3
tokenizers==0.20.3
PyMuPDF
img2pdf
einops
easydict
addict 
Pillow
numpy
pandas
tqdm
transformers >= 4.42.4  # Required for Gemma 2 and for additional chat template parameters.
aiohttp
huggingface_hub[cli]
jinja2
pillow
tenacity